refactor: fix sonarqube warnings
Refs: FILES-581
