+++ src/carbonio-files-ui-common/views/components/sharing/collaborators/EditSharePopoverContainer.tsx
