+++ src/carbonio-files-ui-common/views/components/sharing/collaborators/NewSharePopoverContainer.tsx
